RES version/requirements.txt
absl-py==1.2.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
fonttools==4.34.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
murmurhash==1.0.7
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.1.0
torch==1.12.0+cpu
torch-tb-profiler==0.4.0
torchaudio==0.12.0+cpu
torchvision==0.13.0+cpu
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
Werkzeug==2.2.1
zipp==3.8.1
